+++ ConcernsCaseWork/ConcernsCaseWork.API.Contracts/NtiWarningLetter/CreateNTIWarningLetterRequest.cs
@@ -1,24 +1,22 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace ConcernsCaseWork.API.RequestModels.CaseActions.NTI.WarningLetter
+namespace ConcernsCaseWork.API.Contracts.NtiWarningLetter
 {
-    public class PatchNTIWarningLetterRequest
+    public class CreateNTIWarningLetterRequest
     {
-        [Required]
-        public long Id { get; set; }
-
-        [Required]
+		[Required]
         public int CaseUrn { get; set; }
         public DateTime? DateLetterSent { get; set; }
-		[StringLength(2000)]
-		public string Notes { get; set; }
+        
+        [StringLength(2000)]
+        public string Notes { get; set; }
         public int? StatusId { get; set; }
 
         public ICollection<int> WarningLetterReasonsMapping { get; set; }
         public ICollection<int> WarningLetterConditionsMapping { get; set; }
 
-		[StringLength(300)]
-		public string CreatedBy { get; set; }
+        [StringLength(300)]
+        public string CreatedBy { get; set; }
         public DateTime CreatedAt { get; set; }
         public DateTime? UpdatedAt { get; set; }
         public DateTime? ClosedAt { get; set; }
