+++ ConcernsCaseWork/ConcernsCaseWork.API.Tests/Controllers/NTIWarningLetterControllerTests.cs
@@ -1,5 +1,5 @@
-﻿using ConcernsCaseWork.API.Features.NTIWarningLetter;
-using ConcernsCaseWork.API.RequestModels.CaseActions.NTI.WarningLetter;
+﻿using ConcernsCaseWork.API.Contracts.NtiWarningLetter;
+using ConcernsCaseWork.API.Features.NTIWarningLetter;
 using ConcernsCaseWork.API.ResponseModels;
 using ConcernsCaseWork.API.ResponseModels.CaseActions.NTI.WarningLetter;
 using ConcernsCaseWork.API.UseCases;
