+++ ConcernsCaseWork/ConcernsCaseWork.API.Tests/Controllers/NoticeToImproveControllerTests.cs
@@ -1,20 +1,20 @@
-﻿using ConcernsCaseWork.API.RequestModels.CaseActions.NTI.NoticeToImprove;
+﻿using ConcernsCaseWork.API.Contracts.NoticeToImprove;
+using ConcernsCaseWork.API.Features.NoticeToImprove;
 using ConcernsCaseWork.API.ResponseModels;
 using ConcernsCaseWork.API.ResponseModels.CaseActions.NTI.NoticeToImprove;
 using ConcernsCaseWork.API.UseCases;
 using ConcernsCaseWork.Data.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using FizzWare.NBuilder;
 using FluentAssertions;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using Moq;
+using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 using Xunit;
-using ConcernsCaseWork.API.Features.NoticeToImprove;
 
 namespace ConcernsCaseWork.API.Tests.Controllers
 {
