+++ ConcernsCaseWork/ConcernsCaseWork.API.Tests/Factories/FinancialPlanFactoryTests.cs
@@ -1,5 +1,5 @@
-﻿using ConcernsCaseWork.API.Factories.CaseActionFactories;
-using ConcernsCaseWork.API.RequestModels.CaseActions.FinancialPlan;
+﻿using ConcernsCaseWork.API.Contracts.FinancialPlan;
+using ConcernsCaseWork.API.Factories.CaseActionFactories;
 using ConcernsCaseWork.API.ResponseModels.CaseActions.FinancialPlan;
 using ConcernsCaseWork.Data.Models;
 using FizzWare.NBuilder;
@@ -9,11 +9,8 @@ using Xunit;
 
 namespace ConcernsCaseWork.API.Tests.Factories
 {
-    public class FinancialPlanFactoryTests
+	public class FinancialPlanFactoryTests
     {
-
-
-
         [Fact]
         public void CreateDBModel_ExpectedFinancialPlanCase_WhenCreateFinancialPlanRequestProvided()
         {
