+++ ConcernsCaseWork/ConcernsCaseWork.API/Factories/CaseActionFactories/NTIUnderConsiderationFactory.cs
@@ -1,10 +1,10 @@
-﻿using ConcernsCaseWork.API.RequestModels.CaseActions.NTI.UnderConsideration;
+﻿using ConcernsCaseWork.API.Contracts.NtiUnderConsideration;
 using ConcernsCaseWork.API.ResponseModels.CaseActions.NTI.UnderConsideration;
 using ConcernsCaseWork.Data.Models;
 
 namespace ConcernsCaseWork.API.Factories.CaseActionFactories
 {
-    public static class NTIUnderConsiderationFactory
+	public static class NTIUnderConsiderationFactory
     {
         public static NTIUnderConsideration CreateDBModel(CreateNTIUnderConsiderationRequest createNTIUnderConsiderationRequest)
         {
