+++ ConcernsCaseWork/ConcernsCaseWork.API/Features/FinancialPlan/Create.cs
@@ -1,10 +1,9 @@
-﻿using ConcernsCaseWork.API.Exceptions;
-using ConcernsCaseWork.API.RequestModels.CaseActions.FinancialPlan;
+﻿using ConcernsCaseWork.API.Contracts.FinancialPlan;
+using ConcernsCaseWork.API.Exceptions;
 using ConcernsCaseWork.Data;
 using ConcernsCaseWork.Data.Models;
 using MediatR;
 using Microsoft.EntityFrameworkCore;
-using System.ComponentModel.DataAnnotations;
 
 namespace ConcernsCaseWork.API.Features.FinancialPlan
 {
