+++ ConcernsCaseWork/ConcernsCaseWork.API/Features/NTIWarningLetter/PatchNTIWarningLetter.cs
@@ -1,6 +1,6 @@
-﻿using ConcernsCaseWork.API.Exceptions;
+﻿using ConcernsCaseWork.API.Contracts.NtiWarningLetter;
+using ConcernsCaseWork.API.Exceptions;
 using ConcernsCaseWork.API.Factories.CaseActionFactories;
-using ConcernsCaseWork.API.RequestModels.CaseActions.NTI.WarningLetter;
 using ConcernsCaseWork.API.ResponseModels.CaseActions.NTI.WarningLetter;
 using ConcernsCaseWork.API.UseCases;
 using ConcernsCaseWork.Data.Gateways;
