+++ ConcernsCaseWork/ConcernsCaseWork.API/Features/NoticeToImprove/CreateNoticeToImprove.cs
@@ -1,6 +1,6 @@
-﻿using ConcernsCaseWork.API.Exceptions;
+﻿using ConcernsCaseWork.API.Contracts.NoticeToImprove;
+using ConcernsCaseWork.API.Exceptions;
 using ConcernsCaseWork.API.Factories.CaseActionFactories;
-using ConcernsCaseWork.API.RequestModels.CaseActions.NTI.NoticeToImprove;
 using ConcernsCaseWork.API.ResponseModels.CaseActions.NTI.NoticeToImprove;
 using ConcernsCaseWork.API.UseCases;
 using ConcernsCaseWork.Data.Gateways;
