+++ ConcernsCaseWork/ConcernsCaseWork.API/Features/NoticeToImprove/NoticeToImproveController.cs
@@ -1,4 +1,4 @@
-﻿using ConcernsCaseWork.API.RequestModels.CaseActions.NTI.NoticeToImprove;
+﻿using ConcernsCaseWork.API.Contracts.NoticeToImprove;
 using ConcernsCaseWork.API.ResponseModels;
 using ConcernsCaseWork.API.ResponseModels.CaseActions.NTI.NoticeToImprove;
 using ConcernsCaseWork.API.UseCases;
